+++ POS.Core.Tests/UnitTests/CalculateOrder_CouponSpecs.cs
@@ -15,15 +15,15 @@ namespace POS.Core.Tests
         [TestClass]
         public class WhenSubTotalQualifyForCouponDiscount : SpecsBase<CalculateOrder_Coupon>
         {
-            private decimal _actualResult;
+            private double _actualResult;
             private const int BulkQty = 2;
             private const int FreeQty = 1;
-            private const decimal EligibleAmt = 100m;
-            private const decimal OffAmt = 5m;
+            private const double EligibleAmt = 100d;
+            private const double OffAmt = 5d;
             private const int Quantity = 2;
-            private const decimal TotalAfterBaseDiscount = 105;
+            private const double TotalAfterBaseDiscount = 105;
             private Order _order;
-            private decimal _expectedResult;
+            private double _expectedResult;
 
             protected override void Given()
             {
@@ -61,15 +61,15 @@ namespace POS.Core.Tests
         [TestClass]
         public class WhenSubTotalQualifyForCouponDiscountButNoCoupon : SpecsBase<CalculateOrder_Coupon>
         {
-            private decimal _actualResult;
+            private double _actualResult;
             private const int BulkQty = 2;
             private const int FreeQty = 1;
-            private const decimal EligibleAmt = 100m;
-            private const decimal OffAmt = 5m;
+            private const double EligibleAmt = 100d;
+            private const double OffAmt = 5d;
             private const int Quantity = 2;
-            private const decimal TotalAfterBaseDiscount = 105;
+            private const double TotalAfterBaseDiscount = 105;
             private Order _order;
-            private decimal _expectedResult;
+            private double _expectedResult;
 
             protected override void Given()
             {
@@ -107,15 +107,15 @@ namespace POS.Core.Tests
         [TestClass]
         public class WhenSubTotalNotQualifyForCouponDiscount : SpecsBase<CalculateOrder_Coupon>
         {
-            private decimal _actualResult;
+            private double _actualResult;
             private const int BulkQty = 2;
             private const int FreeQty = 1;
-            private const decimal EligibleAmt = 100m;
-            private const decimal OffAmt = 5m;
+            private const double EligibleAmt = 100d;
+            private const double OffAmt = 5d;
             private const int Quantity = 2;
-            private const decimal TotalAfterBaseDiscount = 95;
+            private const double TotalAfterBaseDiscount = 95;
             private Order _order;
-            private decimal _expectedResult;
+            private double _expectedResult;
 
             protected override void Given()
             {
