+++ POS.Core/Discount/ICalculateDiscount.cs
@@ -2,6 +2,6 @@
 {
     public interface ICalculateDiscount
     {
-        decimal CalculateDiscountPrice();
+        double CalculateDiscountPrice();
     }
 }
